Simplify transformers into single file.
